backend/requirements.txt
# Backend Dependencies for Malware Analysis Sandbox

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Analysis Libraries
yara-python==4.3.1
pefile==2023.2.7
python-magic-bin==0.4.14
ssdeep==3.4

# Database & Cache
pymongo==4.6.1
motor==3.3.2
redis==5.0.1

# Task Queue
celery==5.3.6

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
